AK: Make `SeekableStream::truncate()` take a `size_t`
Similar to the return values earlier, a signed value doesn't really make
sense here. Relying on the much more standard `size_t` makes it easier
to use Stream in all contexts.

+++ AK/BufferedStream.h
@@ -295,7 +295,7 @@ public:
 
         return result;
     }
-    virtual ErrorOr<void> truncate(off_t length) override
+    virtual ErrorOr<void> truncate(size_t length) override
     {
         return m_helper.stream().truncate(length);
     }
